pyproject.toml
[tool.poetry]
name = "emkademy"
version = "0.1.0"
description = ""
authors = ["Kıvanç Yüksel <emkademy@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.11"
pandas = "~=2.0"
matplotlib = "~=3.7"
hydra-core = "~=1.3"
pydantic = "~=1.10"
fsspec = {version = "~=2023.6", extras = ["gcs"]}
gcsfs = "~=2023.6"
torch = "==2.0.0"
lightning = "==2.0.6"
torchmetrics = "~=1.0.1"
transformers = {version = "~=4.31", extras = ["torch"]}
mlflow = "==2.5.0"
psycopg2-binary = "~=2.9"
google-cloud-compute = "~=1.13"

[tool.poetry.group.dev.dependencies]
jupyterlab = "~=4.0"
pytest = "~=7.4"
black = "~=23.7.0"
isort = "~=5.12"
flake8 = "~=6.0"
mypy = "~=1.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120

[tool.isort]
profile = "black"
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 120
indent = 4
lines_between_types = 1
known_first_party = ["cybulde"]
default_section = "THIRDPARTY"
sections = ["FUTURE", "STDLIB", "THIRDPARTY", "FIRSTPARTY", "LOCALFOLDER"]

[tool.mypy]
python_version = "3.10"
show_error_codes = true
ignore_missing_imports = true
install_types = true
non_interactive = true
follow_imports = "silent"
no_strict_optional = true

warn_redundant_casts = false
warn_unused_ignores = true
warn_unused_configs = true
warn_return_any = true
warn_no_return = true
warn_incomplete_stub = true

disallow_subclassing_any = true

disallow_untyped_calls = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
disallow_untyped_decorators = true
check_untyped_defs = true
strict_optional = true

verbosity = 0
